+++ Runtime/Scripts/Timeline/BaseClipData.cs
@@ -11,8 +11,20 @@ internal abstract class BaseClipData : ISerializationCallbackReceiver {
 
 //----------------------------------------------------------------------------------------------------------------------
     #region ISerializationCallbackReceiver
-    public abstract void OnBeforeSerialize();
-    public abstract void OnAfterDeserialize();
+
+    public void OnBeforeSerialize() {
+        m_baseClipDataVersion = CUR_CLIP_DATA_VERSION; 
+            
+        OnBeforeSerializeInternalV();
+    }
+
+    public void OnAfterDeserialize() {
+        OnAfterDeserializeInternalV();
+    }
+    
+    protected abstract void OnBeforeSerializeInternalV();
+    protected abstract void OnAfterDeserializeInternalV();
+    
     #endregion
     
 //----------------------------------------------------------------------------------------------------------------------